requirements.txt
numpy==2.2.4
pygame-ce==2.5.3
